Backend/Voice agent/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "twilio>=9.8.1",
    "websockets>=15.0.1",
]
